# docs/swagger.yaml
basePath: /api
definitions:
  go-marketplace_internal_common.ProblemDetails:
    properties:
      detail:
        description: Specific explanation
        type: string
      errors:
        additionalProperties:
          type: string
        description: Field-specific issues
        type: object
      instance:
        description: Request URI
        type: string
      status:
        description: HTTP status
        type: integer
      title:
        description: Short summary
        type: string
      type:
        description: Error category URI
        type: string
    type: object
  go-marketplace_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  go-marketplace_internal_domain.OrderStatus:
    enum:
    - pending
    - processing
    - packaging
    - shipping
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusProcessing
    - OrderStatusPackaging
    - OrderStatusShipping
    - OrderStatusDelivered
    - OrderStatusCancelled
  go-marketplace_internal_domain.PaymentMethod:
    enum:
    - wallet
    - midtrans
    type: string
    x-enum-varnames:
    - PaymentMethodWallet
    - PaymentMethodMidtrans
...
  internal_core_merchant.MerchantRegisterRequest:
    properties:
      about:
        type: string
      name:
        type: string
      tax_id:
        type: string
    type: object
  internal_core_merchant.MerchantResponse:
    properties:
      about:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  internal_core_order.AppealOrderRequest:
    properties:
      reason:
        type: string
    type: object
  internal_core_order.CheckoutRequest:
    properties:
      address_id:
        type: string
      payment_method:
        $ref: '#/definitions/go-marketplace_internal_domain.PaymentMethod'
      shipping_city:
        type: string
      shipping_phone_number:
        type: string
      shipping_postal_code:
        type: string
      shipping_province:
        type: string
      shipping_recipient_name:
        type: string
      shipping_street_address:
        type: string
    type: object
  internal_core_order.OrderItemResponse:
    properties:
      id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  internal_core_order.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_appealed:
        type: boolean
      items:
        items:
          $ref: '#/definitions/internal_core_order.OrderItemResponse'
        type: array
      merchant_id:
        type: string
      payment_id:
        type: string
      shipping_city:
        type: string
      shipping_phone_number:
        type: string
      shipping_postal_code:
        type: string
      shipping_province:
        type: string
      shipping_recipient_name:
        type: string
      shipping_street_address:
        type: string
      status:
        $ref: '#/definitions/go-marketplace_internal_domain.OrderStatus'
      total_amount:
        type: number
      updated_at:
        type: string
    type: object
  internal_core_order.UpdateStatusRequest:
    properties:
      status:
        $ref: '#/definitions/go-marketplace_internal_domain.OrderStatus'
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a robust e-commerce backend API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Marketplace API
  version: "1.0"
paths:
...
  /auth/register-merchant:
    post:
      consumes:
      - application/json
      description: Allows an authenticated user to register their own shop/merchant
        profile.
      parameters:
      - description: Merchant Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_core_merchant.MerchantRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_core_merchant.MerchantResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Register as merchant
      tags:
      - merchants
...
  /merchants/orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Allows the shop owner to cancel an order and initiate a refund.
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Cancel order (Merchant)
      tags:
      - orders
  /merchants/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Allows the shop owner to change the lifecycle status of an order
        (e.g., to SHIPPED).
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New Status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_core_order.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Update order status (Merchant)
      tags:
      - orders
...
  /users/orders:
    post:
      consumes:
      - application/json
      description: Converts the current items in the authenticated user's cart into
        a formal order.
      parameters:
      - description: Checkout Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_core_order.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_core_order.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Create order from cart
      tags:
      - orders
  /users/orders/{id}:
    get:
      consumes:
      - application/json
      description: Fetches full information for a specific order by its ID.
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_core_order.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get order details
      tags:
      - orders
  /users/orders/{id}/appeal:
    post:
      consumes:
      - application/json
      description: Submits a formal appeal for an order, typically for disputes.
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Appeal Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_core_order.AppealOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Appeal order (User)
      tags:
      - orders
  /users/orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancels a pending or processing order and initiates a refund if
        applicable.
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-marketplace_internal_common.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Cancel order (User)
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"